refactor(cmd): thin init.go; move logic to internal/onboarding
- Move greenfield scaffolding, manifest construction, flavour-detection
  wiring, change-summary diff/apply, and step orchestration into
  internal/onboarding (scaffold.go, manifest_build.go, diff.go,
  output.go, flow.go).
- cmd/wave/commands/init.go reduced from 1575 to 364 LOC, becomes
  cobra wiring + flag parsing + interactive prompts + output
  formatting wrappers.
- CLI behaviour, flags, prompts, and output schemas all unchanged
  (`wave init --help` byte-identical).
- Tests in cmd/wave/commands/init_test.go and forge_init_test.go
  updated to reference the relocated symbols via the onboarding
  package directly.

Partial of #1502 (init.go portion).
